dev-requirements.txt
pytest-ckan
pytest-cov>=2.7.1
pytest-watch
bump2version
